third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-cors-with-credentials-expected.txt

CONSOLE ERROR: Access to resource at 'http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=1' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Origin' header in the response must not be the wildcard '*' when the request's credentials mode is 'include'.
CONSOLE ERROR: Access to resource at 'http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=2' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Origin' header in the response must not be the wildcard '*' when the request's credentials mode is 'include'.
CONSOLE ERROR: Access to resource at 'http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=3' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'.
[Worker] Test that EventSource cross-origin requests with credentials fail until the correct CORS headers are sent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-cors-with-credentials.js
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=1", {'withCredentials': true}) did not throw exception.
PASS [Worker] es.withCredentials is true
PASS [Worker] count != 4 && count != 5 is true
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=2", {'withCredentials': true}) did not throw exception.
PASS [Worker] es.withCredentials is true
PASS [Worker] count != 4 && count != 5 is true
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=3", {'withCredentials': true}) did not throw exception.
PASS [Worker] es.withCredentials is true
PASS [Worker] count != 4 && count != 5 is true
PASS [Worker] es = new EventSource("http://127.0.0.1:8080/eventsource/resources/es-cors-credentials.php?count=4", {'withCredentials': true}) did not throw exception.
PASS [Worker] es.withCredentials is true
PASS [Worker] evt.origin.indexOf('http://127.0.0.1:8080') === 0 is true
PASS [Worker] evt.data is "DATA1"
PASS [Worker] evt.lastEventId is "77"
PASS [Worker] count is 5
PASS [Worker] evt.origin.indexOf('http://127.0.0.1:8080') === 0 is true
PASS [Worker] count is 5
PASS [Worker] evt.data is "DATA2"
PASS successfullyParsed is true

TEST COMPLETE
